add iwyu job

+++ include/chesscxx/core/internal/square.h
@@ -75,7 +75,8 @@ constexpr auto farthestPawnPush(const Square& square, const Color& color)
 }
 
 constexpr auto squareColor(const Square& square) -> Color {
-  auto parity = static_cast<uint>(index(square.file) + index(square.rank)) & 1U;
+  auto parity =
+      static_cast<uint8_t>(index(square.file) + index(square.rank)) & 1U;
   return (parity == 0) ? Color::kWhite : Color::kBlack;
 }
